requirements.txt
pip==22.0.4
wheel==0.37.1
setuptools==62.0.0
fastapi==0.75.0
uvicorn[standard]==0.17.6
google-cloud-datastore==2.5.1
google-auth==2.6.2
python-multipart==0.0.5
pycryptodome==3.14.1
cloud-sql-python-connector[pg8000]==0.6.0
toml-config==0.1.3
pymysql==1.0.2
mysql-connector==2.2.9